feat: update sweep and flip sdk
