+++ packages/sdk/src/sweep-and-flip/types.ts
@@ -1,5 +1,3 @@
-import { TypedDataField } from "ethers";
-
 /**
  * supported swap orders types.
  */
@@ -50,63 +48,4 @@ export type SwapOrderParams = {
    * deadline for the swap in seconds.
    */
   deadline: number;
-};
-
-export const swapOrderTypedData: Record<string, TypedDataField[]> = {
-  SwapOrderParams: [
-    {
-      name: "orderType",
-      type: "OrderType",
-    },
-    {
-      name: "signerAddress",
-      type: "string",
-    },
-    {
-      name: "path",
-      type: "string[]",
-    },
-    {
-      name: "collection",
-      type: "string",
-    },
-    {
-      name: "tokenIds",
-      type: "string[]",
-    },
-    {
-      name: "currency",
-      type: "string",
-    },
-    {
-      name: "amount",
-      type: "string",
-    },
-    {
-      name: "recipient",
-      type: "string",
-    },
-    {
-      name: "deadline",
-      type: "number",
-    },
-  ],
-  OrderType: [
-    {
-      name: "ETH_TO_ERC721",
-      type: "number",
-    },
-    {
-      name: "ERC20_TO_ERC721",
-      type: "number",
-    },
-    {
-      name: "ERC721_TO_ETH",
-      type: "number",
-    },
-    {
-      name: "ERC721_TO_ERC20",
-      type: "number",
-    },
-  ],
 };
